trpc: refactored routing
