+++ src/dictionaries/dictionary.ts
@@ -485,10 +485,10 @@ export interface Dictionary {
             "3": { name: string; description: string };
         },
         statuses: {
-            "1": { name: string; description: string };
-            "2": { name: string; description: string };
-            "3": { name: string; description: string };
-            "4": { name: string; description: string };
+            "1": { name: string; description: string; color: string };
+            "2": { name: string; description: string; color: string };
+            "3": { name: string; description: string; color: string };
+            "4": { name: string; description: string; color: string };
         },
         prios: {
             "0": { name: string; description: string };
