+++ src/server/api/routers/reports/update.ts
@@ -1,110 +1,15 @@
-import { createTRPCRouter, workerProcedure, userProcedure } from "../trpc";
+import { createTRPCRouter, workerProcedure, } from "@/server/api/trpc";
 import { z } from 'zod';
 import { and, eq } from 'drizzle-orm';
-import { db } from "@/server/db";
 import { reports } from "@/server/db/schema/reports";
 import { TRPCError } from "@trpc/server";
-import { pictures } from "@/server/db/schema/pictures";
 import { protocolls } from "@/server/db/schema/protocoll";
 import { users } from "@/server/db/schema/users";
 import { getDictionary } from "@/server/get-dictionary";
 import { sendChangeNotification } from "@/lib/sendChangeNotification";
 import { env } from "@/env";
 
-const STATUS_COLORS = {
-    1: "#4169E1",
-    2: "#FFA500",
-    3: "#32CD32",
-    4: "#FF0000",
-} as const;
-
-export const reportDetailsRouter = createTRPCRouter({
-
-    getWorkerReportDetails: workerProcedure.input(z.object({
-        reportID: z.string(),
-    })).query(async ({ input }) => {
-        const { reportID } = input;
-
-
-        const report = await db.query.reports.findFirst({
-            where: eq(reports.id, parseInt(reportID)),
-        })
-
-        if (!report) {
-            throw new TRPCError({ code: 'NOT_FOUND', message: 'Report not found' });
-        }
-
-        const { prio, ...reportWithoutPriority } = report;
-
-
-        const images = await db.query.pictures.findMany({
-            where: eq(pictures.reportId, report.id),
-        })
-        const imagesProcessed = images.map((image) => ({
-            key: image.id,
-        }))
-
-
-        const protocollData = await db.query.protocolls.findMany({
-            where: eq(protocolls.reportId, report.id),
-        })
-
-        const protocollDataProcessed = protocollData.map((protocoll) => ({
-            timestamp: protocoll.time,
-            status: protocoll.statusId,
-            comment: protocoll.comment,
-            userId: protocoll.userId
-        }))
-
-
-        return {
-            report: { ...reportWithoutPriority, prio },
-            images: imagesProcessed,
-            protocolls: protocollDataProcessed,
-        };
-    }),
-
-    getReportDetails: userProcedure.input(z.object({
-        reportID: z.string(),
-    })).query(async ({ input }) => {
-        const { reportID } = input;
-
-
-        const report = await db.query.reports.findFirst({
-            where: eq(reports.id, parseInt(reportID)),
-        })
-
-        if (!report) {
-            throw new TRPCError({ code: 'NOT_FOUND', message: 'Report not found' });
-        }
-
-        const images = await db.query.pictures.findMany({
-            where: eq(pictures.reportId, report.id),
-        })
-        const imagesProcessed = images.map((image) => ({
-            key: image.id,
-        }))
-
-
-        const protocollData = await db.query.protocolls.findMany({
-            where: eq(protocolls.reportId, report.id),
-        })
-
-        const protocollDataProcessed = protocollData.map((protocoll) => ({
-            timestamp: protocoll.time,
-            status: protocoll.statusId,
-            comment: protocoll.comment,
-            userId: protocoll.userId
-        }))
-
-
-        return {
-            report: report,
-            images: imagesProcessed,
-            protocolls: protocollDataProcessed,
-        };
-    }),
-
+export const reportsUpdateRouter = createTRPCRouter({
     addProtocoll: workerProcedure
         .input(z.object({
             reportId: z.string(),
@@ -162,7 +67,7 @@ export const reportDetailsRouter = createTRPCRouter({
                 firstName: initialUserName ?? "User",
                 title: report.name,
                 status: dictionariesDe.metadata.statuses[statusId as 1 | 2 | 3 | 4].name,
-                status_color: STATUS_COLORS[statusId as 1 | 2 | 3 | 4],
+                status_color: dictionariesDe.metadata.statuses[statusId as 1 | 2 | 3 | 4].color,
                 link: `${env.NEXTAUTH_URL}/myReports/${reportId}`,
                 recipient: initialUserEmail,
                 dictionaries,
@@ -181,7 +86,7 @@ export const reportDetailsRouter = createTRPCRouter({
             return protocol[0];
         }),
 
-    updatePriority: workerProcedure
+    setPriority: workerProcedure
         .input(z.object({
             reportId: z.string(),
             prio: z.number(),
